+++ app/src/main/java/jp/kawagh/kiando/ListScreen.kt
@@ -6,6 +6,7 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.combinedClickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
@@ -13,25 +14,32 @@ import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Menu
 import androidx.compose.material.icons.filled.MoreVert
+import androidx.compose.material.icons.filled.QuestionMark
 import androidx.compose.material.icons.filled.Remove
 import androidx.compose.material.icons.filled.Star
+import androidx.compose.material.icons.filled.Tag
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import jp.kawagh.kiando.models.Tag
 import jp.kawagh.kiando.ui.components.QuestionWithTagsCard
+import jp.kawagh.kiando.ui.components.TagChip
 import jp.kawagh.kiando.ui.theme.BoardColor
 import kotlinx.coroutines.launch
 
 @Preview
 @Composable
 fun PreviewListScreen() {
-    ListScreen(sampleQuestions.map { QuestionWithTags(it, emptyList()) },
-        { _, _ -> {} }, {}, {}, {}, {}, {}, {}, {})
+    ListScreen(
+        QuestionsUiState(
+            sampleQuestions.map { QuestionWithTags(it, emptyList()) }
+        ),
+        { _, _ -> {} }, {}, {}, {}, {}, {}, {}, {}, {}, { _, _ -> {} })
 
 }
 
@@ -40,10 +48,15 @@ sealed class TabItem(val name: String) {
     object Tagged : TabItem("Tagged")
 }
 
+enum class BottomBarItems(val title: String, val icon: ImageVector) {
+    Questions(title = "questions", icon = Icons.Default.QuestionMark),
+    Tags(title = "tags", icon = Icons.Default.Tag),
+}
+
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ListScreen(
-    questions: List<QuestionWithTags>,
+    questionsUiState: QuestionsUiState,
     navigateToQuestion: (Question, Int) -> Unit,
     navigateToDelete: () -> Unit,
     navigateToLicense: () -> Unit,
@@ -52,8 +65,13 @@ fun ListScreen(
     handleFavoriteQuestion: (Question) -> Unit,
     handleInsertSampleQuestions: () -> Unit,
     handleLoadQuestionFromResource: () -> Unit,
+    handleAddTag: (Tag) -> Unit,
+    handleAddCrossRef: (Question, Tag) -> Unit,
 ) {
     var tabRowIndex by remember {
+        mutableStateOf(0)
+    }
+    var bottomBarIndex by remember {
         mutableStateOf(0)
     }
     val navigateToQuestionWithTabIndex: (Question) -> Unit = {
@@ -65,8 +83,8 @@ fun ListScreen(
         mutableStateOf(false)
     }
     val questionsToDisplay = when (tabs[tabRowIndex]) {
-        is TabItem.All -> questions
-        is TabItem.Tagged -> questions.filter { it.question.tag_id != null }
+        is TabItem.All -> questionsUiState.questionsWithTags
+        is TabItem.Tagged -> questionsUiState.questionsWithTags.filter { it.question.tag_id != null }
     }.filter {
         if (hideDefaultQuestions) {
             it.question.id >= 0
@@ -84,6 +102,41 @@ fun ListScreen(
     )
     val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
     val scope = rememberCoroutineScope()
+
+
+    /**
+     * states for tags
+     */
+    var tagNameInput by remember {
+        mutableStateOf("")
+    }
+    var tagDropDownExpanded by remember {
+        mutableStateOf(false)
+    }
+    val tagOptions = questionsUiState.tags
+    var selectedTag by remember {
+        mutableStateOf(
+            if (questionsUiState.tags.isEmpty()) {
+                null
+            } else {
+                Tag(-1, "sample")
+            }
+        )
+    }
+    var questionsDropdownExpanded by remember {
+        mutableStateOf(false)
+    }
+    val questionOptions = questionsUiState.questionsWithTags.map { it.question }
+    var selectedQuestion by remember {
+        mutableStateOf(
+            if (questionsUiState.questionsWithTags.isEmpty()) {
+                null
+            } else {
+                sampleQuestion
+            }
+        )
+    }
+
     ModalNavigationDrawer(
         drawerState = drawerState,
         drawerContent = {
@@ -124,7 +177,14 @@ fun ListScreen(
         }) {
         Scaffold(
             topBar = {
-                TopAppBar(title = { Text(text = stringResource(id = R.string.app_name)) },
+                TopAppBar(title = {
+                    Text(
+                        when (BottomBarItems.values()[bottomBarIndex]) {
+                            BottomBarItems.Questions -> "問題一覧"
+                            BottomBarItems.Tags -> "タグ一覧"
+                        }
+                    )
+                },
                     navigationIcon = {
                         IconButton(onClick = { scope.launch { drawerState.open() } }) {
                             Icon(Icons.Default.Menu, null)
@@ -141,47 +201,203 @@ fun ListScreen(
                     }
                 )
             },
+            bottomBar = {
+                NavigationBar {
+                    BottomBarItems.values().forEachIndexed { index, bottomBarItem ->
+                        NavigationBarItem(
+                            selected = bottomBarIndex == index,
+                            onClick = { bottomBarIndex = index },
+                            label = { Text(bottomBarItem.title) },
+                            icon = { Icon(bottomBarItem.icon, null) })
+
+                    }
+                }
+            }
         ) { paddingValues ->
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(paddingValues),
-                horizontalAlignment = Alignment.CenterHorizontally,
-            ) {
-                TabRow(selectedTabIndex = tabRowIndex) {
-                    tabs.forEachIndexed { index, tab ->
-                        Tab(
-                            selected = tabRowIndex == index,
-                            onClick = { tabRowIndex = index }) {
-                            Text(
-                                tab.name,
-                                fontSize = MaterialTheme.typography.titleLarge.fontSize,
+            when (BottomBarItems.values()[bottomBarIndex]) {
+                BottomBarItems.Questions -> {
+                    Column(
+                        modifier = Modifier
+                            .fillMaxSize()
+                            .padding(paddingValues),
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                    ) {
+                        TabRow(selectedTabIndex = tabRowIndex) {
+                            tabs.forEachIndexed { index, tab ->
+                                Tab(
+                                    selected = tabRowIndex == index,
+                                    onClick = { tabRowIndex = index }) {
+                                    Text(
+                                        tab.name,
+                                        fontSize = MaterialTheme.typography.titleLarge.fontSize,
+                                    )
+                                }
+                            }
+                        }
+                        Text(
+                            text = "Problem Set",
+                            fontSize = MaterialTheme.typography.headlineSmall.fontSize
+                        )
+                        if (questionsUiState.questionsWithTags.isEmpty()) {
+                            Box(Modifier.fillMaxSize()) {
+                                Column(
+                                    modifier = Modifier.align(Alignment.Center),
+                                    horizontalAlignment = Alignment.CenterHorizontally,
+                                ) {
+                                    Text(
+                                        "no questions",
+                                        style = MaterialTheme.typography.headlineSmall
+                                    )
+                                    Button(onClick = handleInsertSampleQuestions) { Text("add sample questions") }
+                                }
+                            }
+                        } else {
+                            QuestionsList(
+                                questionsWithTags = questionsToDisplay,
+                                navigateToQuestion = navigateToQuestionWithTabIndex,
+                                handleDeleteAQuestion = handleDeleteAQuestion,
+                                handleRenameAQuestion = handleRenameAQuestion,
+                                handleFavoriteQuestion = handleFavoriteQuestion,
                             )
                         }
                     }
+
                 }
-                Text(
-                    text = "Problem Set",
-                    fontSize = MaterialTheme.typography.headlineSmall.fontSize
-                )
-                if (questions.isEmpty()) {
-                    Box(Modifier.fillMaxSize()) {
-                        Column(
-                            modifier = Modifier.align(Alignment.Center),
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                        ) {
-                            Text("no questions", style = MaterialTheme.typography.headlineSmall)
-                            Button(onClick = handleInsertSampleQuestions) { Text("add sample questions") }
+
+                BottomBarItems.Tags -> {
+                    LazyColumn(
+                        Modifier
+                            .padding(paddingValues)
+                            .padding(horizontal = 16.dp)
+                    ) {
+                        item {
+                            OutlinedTextField(
+                                value = tagNameInput,
+                                onValueChange = { tagNameInput = it },
+                                label = { Text("tag") },
+                                trailingIcon = {
+                                    IconButton(onClick = {
+                                        if (tagNameInput.isNotEmpty()) {
+                                            handleAddTag(Tag(title = tagNameInput))
+                                        }
+                                    }) {
+                                        Icon(Icons.Default.Add, null)
+                                    }
+                                },
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                            )
+
+                        }
+
+                        item { Text("tags") }
+
+                        item {
+                            LazyRow(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
+                                items(questionsUiState.tags) {
+                                    TagChip(tag = it)
+                                }
+                            }
+
+                        }
+
+                        item {
+                            Divider()
+                        }
+
+                        item {
+                            ExposedDropdownMenuBox(
+                                expanded = questionsDropdownExpanded,
+                                onExpandedChange = {
+                                    questionsDropdownExpanded = !questionsDropdownExpanded
+                                }) {
+                                OutlinedTextField(
+                                    value = selectedQuestion?.description ?: "問題を選択してください",
+                                    onValueChange = {},
+                                    label = { Text("問題") },
+                                    trailingIcon = {
+                                        ExposedDropdownMenuDefaults.TrailingIcon(
+                                            expanded = questionsDropdownExpanded
+                                        )
+                                    },
+                                    readOnly = true,
+                                    modifier = Modifier
+                                        .menuAnchor()
+                                        .fillMaxWidth()
+                                )
+                                ExposedDropdownMenu(
+                                    expanded = questionsDropdownExpanded,
+                                    onDismissRequest = { questionsDropdownExpanded = false }) {
+                                    questionOptions.forEach {
+                                        DropdownMenuItem(
+                                            text = { Text(it.description) },
+                                            onClick = {
+                                                selectedQuestion = it
+                                                questionsDropdownExpanded = false
+                                            },
+                                            contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding
+                                        )
+                                    }
+                                }
+
+                            }
+                        }
+
+                        item {
+                            ExposedDropdownMenuBox(
+                                expanded = tagDropDownExpanded,
+                                onExpandedChange = {
+                                    tagDropDownExpanded = !tagDropDownExpanded
+                                }) {
+                                OutlinedTextField(
+                                    value = selectedTag?.title ?: "タグを選択してください",
+                                    onValueChange = {},
+                                    label = { Text("タグ") },
+                                    trailingIcon = {
+                                        ExposedDropdownMenuDefaults.TrailingIcon(
+                                            expanded = tagDropDownExpanded
+                                        )
+                                    },
+                                    readOnly = true,
+                                    modifier = Modifier
+                                        .menuAnchor()
+                                        .fillMaxWidth()
+                                )
+                                ExposedDropdownMenu(
+                                    expanded = tagDropDownExpanded,
+                                    onDismissRequest = { tagDropDownExpanded = false }) {
+                                    tagOptions.forEach {
+                                        DropdownMenuItem(
+                                            text = { Text(it.title) },
+                                            onClick = {
+                                                selectedTag = it
+                                                tagDropDownExpanded = false
+                                            },
+                                            contentPadding = ExposedDropdownMenuDefaults.ItemContentPadding
+                                        )
+                                    }
+                                }
+                            }
+                        }
+
+                        item {
+                            Column(
+                                horizontalAlignment = Alignment.CenterHorizontally,
+                                modifier = Modifier.fillMaxWidth()
+                            ) {
+                                Button(onClick = {
+                                    selectedQuestion?.let { question ->
+                                        selectedTag?.let { tag ->
+                                            handleAddCrossRef(question, tag)
+                                        }
+                                    }
+                                })
+                                {
+                                    Text("問題にタグを追加")
+                                }
+                            }
                         }
                     }
-                } else {
-                    QuestionsList(
-                        questionsWithTags = questionsToDisplay,
-                        navigateToQuestion = navigateToQuestionWithTabIndex,
-                        handleDeleteAQuestion = handleDeleteAQuestion,
-                        handleRenameAQuestion = handleRenameAQuestion,
-                        handleFavoriteQuestion = handleFavoriteQuestion,
-                    )
                 }
             }
         }