+++ app/src/main/java/jp/kawagh/kiando/QuestionsViewModel.kt
@@ -31,6 +31,11 @@ class QuestionsViewModel @Inject constructor(
         viewModelScope.launch() {
             db.questionDao().getQuestionsWithTags().collectLatest {
                 uiState = uiState.copy(questionsWithTags = it)
+            }
+        }
+        viewModelScope.launch {
+            db.tagDao().getAll().collect {
+                uiState = uiState.copy(tags = it)
 
             }
         }
@@ -85,6 +90,18 @@ class QuestionsViewModel @Inject constructor(
         }
     }
 
+    fun add(tag: Tag) {
+        viewModelScope.launch(Dispatchers.IO) {
+            db.tagDao().insert(tag)
+        }
+    }
+
+    fun addCrossRef(question: Question, tag: Tag) {
+        viewModelScope.launch(Dispatchers.IO) {
+            db.questionTagCrossRefDao().insert(QuestionTagCrossRef(question.id, tag.id))
+        }
+    }
+
     fun loadQuestionsFromAsset() {
         val csvQuestionStream = context.resources.openRawResource(R.raw.questions)
         val csvReader = csvReader { this.delimiter = ';' }
@@ -105,5 +122,6 @@ class QuestionsViewModel @Inject constructor(
 }
 
 data class QuestionsUiState(
-    val questionsWithTags: List<QuestionWithTags> = emptyList()
+    val questionsWithTags: List<QuestionWithTags> = emptyList(),
+    val tags: List<Tag> = emptyList()
 )